+++ lib/index.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="react" />
 import { Props } from './interfaces';
 declare const NativeColorPicker: ({ animate, colors, columns, gradient, horizontal, itemSize, markerType, markerDisplay, onSelect, selectedColor, shadow, sort, itemProps, itemStyle, linearGradientProps, linearGradientStyle, ...props }: Props) => JSX.Element;
 export default NativeColorPicker;